Remove Monsters.java
incorporatedMonster.java into Registry and Monster
