+++ src/New/Monster.java
@@ -23,4 +23,15 @@ public class Monster {
 			AllMonsters.add(this);
 		}
 	}
+	public Monster pickMonster(Player a){
+		Random rand = new Random();
+		Monster m = new Monster("Read Only", 0, 0, 0, 0, -1);
+		while(m.lvl > a.lvl){
+			//while required level is greater then player level, pick new monster
+			//this is so we pick a monster within the meetable monsters for this player
+			m = m.AllMonsters.get(rand.nextInt(m.AllMonsters.size()));
+			//this picks a random monster within AllMonsters
+		}
+		return m;
+	}
 }
